+++ app/Http/Controllers/GalerijaController.php
@@ -8,6 +8,8 @@ use Image;
 use File;
 use Illuminate\Http\Request;
 use Mail;
+use Illuminate\Support\Facades\Auth;
+
 
 class GalerijaController extends Controller
 {
@@ -20,62 +22,61 @@ class GalerijaController extends Controller
     }
     public function update(Request $request, $id)
     {
-        $galerija=  Galerija::find($id);
         $request->validate([
             "naziv"=>'required|max:30',
             'file' => 'image|mimes:jpeg,bmp,png' 
 
         ]);
-        if($galerija==null){
-            return back();
-        }
-       
+        if(Auth::check()){
+            if(auth()->user()->hasRole('admin')){
+                $galerija=  Galerija::find($id);
 
-        $imagedb=null;
-        // ensure the request has a file before we attempt anything else.
-        if ($request->hasFile('file')) {
-            $request->validate([
-                'file' => 'required|image|mimes:jpeg,bmp,png' 
+                
 
-    
-            ]);
+                $imagedb=null;
+                if ($request->hasFile('file')) {
+                
+                    
+                    $image = $request->file('file');
+                    $input['imagename'] = time().'.'.$image->extension();
             
-            $image = $request->file('file');
-             $input['imagename'] = time().'.'.$image->extension();
-     
-            $filePath = public_path('/images');
+                    $filePath = public_path('/images');
 
 
-            $img = Image::make($image->path());
-            $img->resize(430, 720, function ($const) {
-                $const->aspectRatio();
-                $const->upsize();
-            })->save($filePath.'/'.$input['imagename']);
+                    $img = Image::make($image->path());
+                    $img->resize(430, 720, function ($const) {
+                        $const->aspectRatio();
+                        $const->upsize();
+                    })->save($filePath.'/'.$input['imagename']);
 
-            Images::create([
-                "name" => $input['imagename'],
-                "file_path" =>  $filePath]);
-           
-            $imagedb= Images::get()->where( 'name', '=', $input['imagename'])->first();
+                    Images::create([
+                        "name" => $input['imagename'],
+                        "file_path" =>  $filePath]);
+                
+                    $imagedb= Images::get()->where( 'name', '=', $input['imagename'])->first();
+                }
+
+                $galerija->name=$request->get('naziv');
+                $galerija->save();
+                if($imagedb!=null){
+
+                    $image=Images::get()->find($galerija->images_id);
+                    $galerija->images_id=$imagedb->id;
+                    $galerija->save();
+                    $filename=$image->file_path.'/'.$image->name;
+                    File::delete($filename);
+                    $image->delete();
+                }
+            
+                $request->session()->flash('alert-success', 'Uspješno izmjenjena slika.');
+            
+                return redirect()->route('galerija');
+            }
         }
+        $request->session()->flash('alert-warning', 'Za to akciju nemate privilegije.');
 
-         
-          $galerija->name=$request->get('name');
-          $galerija->save();
-          if($imagedb!=null){
-
-            $image=Images::get()->find($galerija->images_id);
-            $galerija->images_id=$imagedb->id;
-            $galerija->save();
-            $filename=$image->file_path.'/'.$image->name;
-            File::delete($filename);
-            $image->delete();
-          }
-
-      
-          $request->session()->flash('alert-success', 'Uspjesno izmjenjena galerija.');
-    
         return redirect()->route('galerija');
+
     }
     public function show(Galerija $galerija){
         return view('galerija.galerijaUpdate',['galerija'=>$galerija]);
@@ -84,62 +85,72 @@ class GalerijaController extends Controller
     {
         // Validate the inputs
         $request->validate([
-            "naziv"=>'required|max:30'
+            "naziv"=>'required|max:30',
             'file' => 'required|image|mimes:jpeg,bmp,png' 
 
         ]);
+        if(Auth::check()){
+            if(auth()->user()->hasRole('admin')){
+                if ($request->hasFile('file')) {
 
-        // ensure the request has a file before we attempt anything else.
-        if ($request->hasFile('file')) {
-
+                    
+                    $image = $request->file('file');
+                
+                    $input['imagename'] = time().'.'.$image->extension();
             
-            $image = $request->file('file');
-           
-             $input['imagename'] = time().'.'.$image->extension();
-     
-            $filePath = public_path('/images');
+                    $filePath = public_path('/images');
 
 
-            $img = Image::make($image->path());
-            $img->resize(430, 720, function ($const) {
-                $const->aspectRatio();
-                $const->upsize();
-            })->save($filePath.'/'.$input['imagename']);
+                    $img = Image::make($image->path());
+                    $img->resize(430, 720, function ($const) {
+                        $const->aspectRatio();
+                        $const->upsize();
+                    })->save($filePath.'/'.$input['imagename']);
 
-            Images::create([
-                "name" => $input['imagename'],
-                "file_path" =>  $filePath]);
-            /* 
-            https://www.positronx.io/how-to-resize-images-in-laravel-before-uploading/
+                    Images::create([
+                        "name" => $input['imagename'],
+                        "file_path" =>  $filePath]);
+                    /* 
+                    https://www.positronx.io/how-to-resize-images-in-laravel-before-uploading/
+                    
+                    redi ako se kreira link storage i public 
+                    $request->file->store('slike', 'public');
+                    Images::create([
+                            "name" =>$request->get('name'),
+                            "file_path" => $request->file->hashName()
+                    ]);
+                    $imagedb= Images::get()->where('file_path','=',$request->file->hashName())->first();*/
+
+                    $imagedb= Images::get()->where( 'name', '=', $input['imagename'])->first();
+                    Galerija::create([
+                            "name" =>$request->get('naziv'),
+                            "images_id" => $imagedb->id,
+                            "kreirao_id" =>auth()->id()]);
+                }
+                $request->session()->flash('alert-success', 'Ušpjesno dodana slika.');
             
-            redi ako se kreira link storage i public 
-            $request->file->store('slike', 'public');
-            Images::create([
-                    "name" =>$request->get('name'),
-                    "file_path" => $request->file->hashName()
-            ]);
-            $imagedb= Images::get()->where('file_path','=',$request->file->hashName())->first();*/
+                return back();
+            }
+        }   
+        $request->session()->flash('alert-warning', 'Za to akciju nemate privilegije.');
 
-            $imagedb= Images::get()->where( 'name', '=', $input['imagename'])->first();
-            Galerija::create([
-                    "name" =>$request->get('naziv'),
-                    "images_id" => $imagedb->id,
-                    "kreirao_id" =>auth()->id()]);
-        }
-        $request->session()->flash('alert-success', 'Uspjesno dodana slika.');
-    
         return back();
+        
     }
     public function destroy(Galerija $galerija){
         
-          $image=Images::get()->find($galerija->images_id);
-          $galerija->delete();
-          $filename=$image->file_path.'/'.$image->name;
-          File::delete($filename);
-         //unlink($filename);
-          $image->delete();   
-          $request->session()->flash('alert-success', 'Uspjesno izbrisana slika.');
+        if(Auth::check()){
+            if(auth()->user()->hasRole('admin')){
+                $image=Images::get()->find($galerija->images_id);
+                $galerija->delete();
+                $filename=$image->file_path.'/'.$image->name;
+                File::delete($filename);
+                //unlink($filename);
+                $image->delete();   
 
+            }
+        } 
         return back();
+
     }
 }
